switch: arizona: add software debounce for antenna
When antenna cable is plugged in we continously measure the impedance
of micdet pin so as to detect a subsequent plugin of 3/4 pole jack.
In such a scenario if a 4 pole headset jack is slowly plugged in then
hpl/hpr may come in contact with micdet pin and we may meausre the
impedance of hpl/r and falsely report a 4 pole headset as a 3 pole
headphone jack. So this patch adds a software debounce to negate the
effect of slow jack insertions into an antenna cable.

Change-Id: Ib41eb22383cd755659439ee2513400dfa64df1f9
Signed-off-by: Nikesh Oswal <Nikesh.Oswal@wolfsonmicro.com>