switch: arizona: debounce for jack plugouts from ant cable
a slow removal of 4 pole headset jack from an antenna cable can
result in the ground pin of headset jack to comes in contact with
the mic pin of antenna and due to zero impedance of ground pin the
state machine will falsely report a button press this patch prevents
this situation by adding a software debounce in the state machine.
Also plugin and plugout debounce variables are kept separate so
that each can be tuned according to the requirements.

Change-Id: I3c76678c062b812f82f13117a3987e1b464ba57e
Signed-off-by: Nikesh Oswal <Nikesh.Oswal@wolfsonmicro.com>