ASoC: arizona: Simplify ez2control trigger callback
Now we have a power event for the ADSP itself we can simply clear the
trigger flag every time we boot the ADSP which makes the whole process
much simpler. Additionally, this fixes a race condition where if the
stream is closed the stream free callback would run resetting the
trigger flag before the DSP was powered down. Meaning an IRQ in that
time would be treated as a new trigger.

Change-Id: Ic3041a05ece7ba9159cf6111b6c196baa931a855
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>